Generate new grid on user input
The "Create new grid" button sends the user a popup asking for the numbe
r of squares per side of the new grid and generates a new grid with the
specified size in the same space that was occupied by the previous grid.
The max number of squares per side of the new grid is 100.